+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Joyride
 ## [Unreleased]
 
 - Add `who` badge and rework info line in terminal output
+- Log agent evaluations to Calva's output subscriber bus when Calva is installed
 
 ## [0.0.72] - 2025-12-20